backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
pyotp
qrcode
python-multipart
pytest
pillow